+++ client/src/Components/Home/Home.jsx
@@ -5,11 +5,25 @@ import Footer from "../Footer/Footer"
 import imagen1 from '../../Assets/imagen1.jpg';
 import imagen2 from '../../Assets/imagen2.jpg';
 import imagen3 from '../../Assets/imagen3.jpg';
+import { useNavigate } from 'react-router-dom';
+
 
 
 export default function Home() {
+  const navigate = useNavigate();
+
+  const handleShowUsers = () =>{
+    navigate('/users');
+  }
+
+  const handleShowDoctors = () =>{
+    navigate('/doctors');
+  }
+
     return (
       <div>
+      <button onClick={handleShowUsers}>Usuarios</button>
+      <button onClick={handleShowDoctors}>Doctores</button>
         <div>
           <NavBar/>
         </div>
