+++ client/src/Doctors/Doctors.jsx
@@ -2,9 +2,9 @@ import React, { useEffect, useState } from "react";
 import { useSelector, useDispatch } from "react-redux";
 import { getDoctors } from "../Redux/actions";
 import Doctor from './Doctor';
-import SearchBar from './SearchBar';
+import SearchBar from '../Components/NavBar/SearchBar';
 import { Link } from 'react-router-dom';
-import Paged from "./Paged";
+import Paged from "../Components/Paged";
 import './doctors.css';
